elapsedtime(mtsdi)	R Documentation

Elapsed Time

Description

Compute the elapsed time between start time and end time

Usage

elapsedtime(st, et)

Arguments

st 	 starting time

et 	 ending time

Details

It returns the time the process took to run.

Value

String of the form hh:mm:ss

Note

It is not intended to be called directly.

Author(s)

Washington Leite Junger wjunger@ims.uerj.br and Antonio Ponce de Leon ponce@ims.uerj.br

See Also

mnimput

[Package mtsdi version 0.3.0 Index]