getmean(mtsdi)	R Documentation

Rows Mean Estimation

Description

Estimate the row mean from a mtsdi object regarding a fixed number of imputed values

Usage

getmean(object, weighted=TRUE, mincol=1, maxconsec=3)

Arguments

object 	 imputation object

weighted 	 If TRUE, weights returned by mnimput will be used form mean computation

mincol 	 integer for the minimun number of valid values by row

maxconsec 	 integer for the maximum number of consecutive missing values in a column

Details

It is useful just in case one wants row mean estimated. If log tranformation was used, mean is adjusted accordingly.

Value

A vector of of rows mean with lenght n, where n is the number of observations.

Author(s)

Washington Leite Junger wjunger@ims.uerj.br and Antonio Ponce de Leon ponce@ims.uerj.br

See Also

mnimput, getmean, edaprep

Examples

data(miss)
f <- ~c31+c32+c33+c34+c35
i <- mnimput(f,miss,eps=1e-3,ts=TRUE, method="spline",sp.control=list(df=c(7,7,7,7,7)))
m <- getmean(i,2)

[Package mtsdi version 0.3.0 Index]