getvalues(mtsdi)	R Documentation

imputed dataset extraction

Description

Extract imputed dataset from a mtsdi object

Usage

getvalues(object)

Arguments

object 	 imputation object

Details

If log tranformation was used, dataset is adjusted accordingly.

Value

A vector of of rows mean with lenght n, where n is the number of observations.

Author(s)

Washington Leite Junger
wjunger@ims.uerj.br

References

Junger, W. L. (2002) Imputação de dados faltando em séries temporais multivariadas via algoritmo EM. Monografia. IME/UERJ - Rio de Janeiro.

Johnson, R., Wichern, D. (1998) Applied Multivariate Statistical Analysis. New Jersey : Prentice Hall.

See Also

mnimput, getmean, edaprep

Examples

library(mtsdi)
d <- mkjnw()
i <- mnimput(~X1+X2+X3,d)
getmean(i,2)

[Package mtsdi version 0.3.0 Index]