plot.mtsdi(mtsdi)	R Documentation

Plot an Imputed Matrix

Description

This function produces a plot with imputed values and the estimated level for each of the columns in the imputed matrix.

Usage

## S3 method for class 'mtsdi':
plot(x, vars = "all", overlay = TRUE, level = TRUE, 
        points = FALSE, leg.loc = "topright", horiz = FALSE, at.once = FALSE, ...)

Arguments

x 	 an object of the class mtsdi

vars 	 a vector with de variables to plot

overlay 	 logical. If TRUE, observed values are plot over the imputed ones

level 	 logical. If TRUE, the level is plot

points 	 logical. If TRUE, points on the observed and imputed values are plot

leg.loc 	 a list with x and y coordinates for the legend or a quoted string. Default is "topright". See Details

horiz 	 logical. If TRUE, the legend will horizontal oriented

at.once 	 logical. If TRUE, all the variables are plot in separate windows at once

... 	 further options for function plot

Details

The leg.loc option may also be specified by setting one of the following quoted strings "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right", or "center". This places the legend on the inside of the plot frame at the given location with the orietation set by horiz. See legend for further details.

Author(s)

Washington Leite Junger wjunger@ims.uerj.br and Antonio Ponce de Leon ponce@ims.uerj.br

See Also

mnimput

Examples

data(miss)
f <- ~c31+c32+c33+c34+c35
i <- mnimput(f,miss,eps=1e-3,ts=TRUE, method="spline",sp.control=list(df=c(7,7,7,7,7)))
plot(i)

[Package mtsdi version 0.3.0 Index]