predict.mtsdi(mtsdi)	R Documentation

Imputed Dataset Extraction

Description

Extract imputed dataset from a mtsdi object

Usage

## S3 method for class 'mtsdi':
predict(object, ...)

Arguments

object 	 imputation object

... 	 further options passed to the generic function predict

Details

If log tranformation was used, dataset is back transformed accordingly.

Value

A vector of of rows mean with lenght n, where n is the number of observations.

Author(s)

Washington Leite Junger wjunger@ims.uerj.br and Antonio Ponce de Leon ponce@ims.uerj.br

References

Junger, W. L. Ponce de Leon, A. Santos, N. (2003) Missing Data Imputation in Multivariate Time Series via EM Algorithm. Cadernos do IME 15, 8–21.

Johnson, R., Wichern, D. (1998) Applied Multivariate Statistical Analysis. Prentice Hall.

Dempster, A., Laird, N., Rubin, D. (1977) Maximum Likelihood from Incomplete Data via the Algorithm EM. Journal of the Royal Statistical Society 39(B)), 1–38.

McLachlan, G. J., Krishnan, T. (1997) The EM algorithm and extensions. John Wiley and Sons.

Box, G., Jenkins, G., Reinsel, G. (1994) Time Series Analysis: Forecasting and Control. 3 ed. Prentice Hall.

Hastie, T. J.; Tibshirani, R. J. (1990) Generalized Additive Models. Chapman and Hall.

See Also

mnimput, getmean, edaprep

Examples

data(miss)
f <- ~c31+c32+c33+c34+c35
i <- mnimput(f,miss,eps=1e-3,ts=TRUE, method="spline",sp.control=list(df=c(7,7,7,7,7)))
predict(i)

[Package mtsdi version 0.3.0 Index]