summary.mtsdi(mtsdi)	R Documentation

Summary Information

Description

Print summary information on the imputation object

Usage

summary.mtsdi(object, ...)

Arguments

object 	 an object of class mtsdi

... 	 further options passed to print.summary.mtsdi

Details

Value

The function resturns a list containing

call 	 function call

muhat 	 estimated mean vector

sigmahat 	 estimated covariance matrix

iterations 	 number of iterations used

convergence 	 relative difference of covariance determinant reached

time 	 time used in the process

models 	 details on the models used for time filtering

log 	 a logical indicating that data are log transformed

log.offset 	 offset used in the log transformation in order to avoid zeros

Author(s)

Washington Leite Junger wjunger@ims.uerj.br and Antonio Ponce de Leon ponce@ims.uerj.br

References

Junger, W. L. Ponce de Leon, A. Santos, N. (2003) Missing Data Imputation in Multivariate Time Series via EM Algorithm. Cadernos do IME 15, 8–21.

Johnson, R., Wichern, D. (1998) Applied Multivariate Statistical Analysis. Prentice Hall.

Dempster, A., Laird, N., Rubin, D. (1977) Maximum Likelihood from Incomplete Data via the Algorithm EM. Journal of the Royal Statistical Society 39(B)), 1–38.

McLachlan, G. J., Krishnan, T. (1997) The EM algorithm and extensions. John Wiley and Sons.

Box, G., Jenkins, G., Reinsel, G. (1994) Time Series Analysis: Forecasting and Control. 3 ed. Prentice Hall.

Hastie, T. J.; Tibshirani, R. J. (1990) Generalized Additive Models. Chapman and Hall.

See Also

mnimput, predict

Examples

data(miss)
f <- ~c31+c32+c33+c34+c35
i <- mnimput(f,miss,eps=1e-3,ts=TRUE, method="spline",sp.control=list(df=c(7,7,7,7,7)))
summary(i)

[Package mtsdi version 0.3.0 Index]